+++ src/learning-services/lingualeo.ts
@@ -1,5 +1,10 @@
 class Lingualeo {
-  public static addWord(word: string, translate: string) {
+  public color: string;
+
+  constructor() {
+    this.color = "#FFC900";
+  }
+  public addWord(word: string, translate: string, _: string) {
     const url = new URL("https://api.lingualeo.com/addword");
     const params: any = { word: word, tword: translate };
     Object.keys(params).forEach(key => url.searchParams.append(key, params[key]));